C version of stb_image
